requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy<2.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=9.0.0

# Screen Capture
mss>=6.1.0

# Input Devices
evdev>=1.6.0
pynput>=1.7.6

# Data Processing
pandas>=1.5.0
pyarrow>=10.0.0  # For parquet support

# Configuration
PyYAML>=6.0

# Optional: ROS2 (for real-time state data)
# Uncomment if using ROS2 bridge
# rclpy

# Development
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0
